Moves parser API to its own directory
Moves the parser API to its own directory to follow the project architecture and improve separation of concerns.

This includes moving the Netlify function, the netlify.toml configuration, and the package.json and package-lock.json files to the new directory.

Also, this commit adds CORS headers to the parser function to allow cross-origin requests.